checks cli utility locaitons
checks to see that the utilities necessary to run the scripts in the microservices suite are installed correctly. if they're not, returns an errorlevel of 1 and exits the script, displaying a warning not to continue. this is a workaround because idk how to push the exit command back to the main shell yet